feat(wip): allow finch to build on Linux without Lima
