+++ cmd/finch/nerdctl_remote.go
@@ -1,0 +1,36 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+//go:build (darwin || windows) && !native
+
+package main
+
+import (
+	"fmt"
+
+	"github.com/runfinch/finch/pkg/command"
+	"github.com/runfinch/finch/pkg/flog"
+	"github.com/runfinch/finch/pkg/lima"
+)
+
+func (nc *nerdctlCommand) assertVMIsRunning(creator command.LimaCmdCreator, logger flog.Logger) error {
+	// Extra call to check VM before running nerdctl commands. These are the reasons of not doing message replacing
+	// 1. for the non-help commands, replacing stdout may cause "stdin is not a terminal" error for the commands that need input.
+	// E.g. finch login.
+	// 2. an extra call could give us more control about the error messages. Message replacing may fail if upstream
+	// changes the format of source string, which leads to extra CI validation work.
+	status, err := lima.GetVMStatus(creator, logger, limaInstanceName)
+	if err != nil {
+		return err
+	}
+	switch status {
+	case lima.Nonexistent:
+		return fmt.Errorf("instance %q does not exist, run `finch %s init` to create a new instance",
+			limaInstanceName, virtualMachineRootCmd)
+	case lima.Stopped:
+		return fmt.Errorf("instance %q is stopped, run `finch %s start` to start the instance",
+			limaInstanceName, virtualMachineRootCmd)
+	default:
+		return nil
+	}
+}
