+++ cmd/finch/nerdctl_windows.go
@@ -15,6 +15,20 @@ import (
 	"github.com/runfinch/finch/pkg/flog"
 )
 
+func (nc *nerdctlCommand) GetCmdArgs() []string {
+	wd, err := nc.systemDeps.GetWd()
+	if err != nil {
+		nc.logger.Warnln("failed to get working directory, will default to user home with error %s", err.Error())
+		return []string{"shell", limaInstanceName, "sudo", "-E"}
+	}
+	wslPath, err := convertToWSLPath(nc.systemDeps, wd)
+	if err != nil {
+		nc.logger.Warnln("failed to convert to WSL path, will default to user home with error %s", err.Error())
+		return []string{"shell", limaInstanceName, "sudo", "-E"}
+	}
+	return []string{"shell", "--workdir", wslPath, limaInstanceName, "sudo", "-E"}
+}
+
 func convertToWSLPath(systemDeps NerdctlCommandSystemDeps, winPath string) (string, error) {
 	path, err := systemDeps.FilePathAbs(filepath.Clean(winPath))
 	if err != nil {
