+++ cmd/finch/support_bundle_native.go
@@ -1,10 +1,10 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build darwin
+//go:build linux || native
 
 package main
 
-func (nc *nerdctlCommand) GetLimaArgs() []string {
-	return []string{"shell", limaInstanceName, "sudo", "-E"}
+func (gsa *generateSupportBundleAction) assertVMExists() error {
+	return nil
 }
