+++ cmd/finch/version.go
@@ -5,8 +5,6 @@ package main
 
 import (
 	"bytes"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"text/template"
@@ -15,7 +13,6 @@ import (
 
 	"github.com/runfinch/finch/pkg/command"
 	"github.com/runfinch/finch/pkg/flog"
-	"github.com/runfinch/finch/pkg/lima"
 	"github.com/runfinch/finch/pkg/templates"
 	"github.com/runfinch/finch/pkg/version"
 )
@@ -165,40 +162,6 @@ func (va *versionAction) showVersionMessage(tmpl *template.Template, nerdctlVers
 		return err
 	}
 	if _, err := fmt.Fprintln(va.stdOut, b.String()); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (va *versionAction) printVersion(format string) error {
-	status, err := lima.GetVMStatus(va.creator, va.logger, limaInstanceName)
-	if err != nil {
-		return fmt.Errorf("failed to get VM status: %w", err)
-	}
-	if status != lima.Running {
-		return errors.New("detailed version info is unavailable because VM is not running")
-	}
-	// Add -E to sudo command in order to preserve existing environment variables, more info:
-	// https://stackoverflow.com/questions/8633461/how-to-keep-environment-variables-when-using-sudo/8633575#8633575
-	limaArgs := []string{"shell", limaInstanceName, "sudo", "-E", "nerdctl", "version", "--format", "json"}
-	out, err := va.creator.CreateWithoutStdio(limaArgs...).Output()
-	if err != nil {
-		return fmt.Errorf("failed to create the nerdctl version command: %w", err)
-	}
-
-	var nerdctlVersion NerdctlVersionOutput
-	err = json.Unmarshal(out, &nerdctlVersion)
-	if err != nil {
-		return fmt.Errorf("failed to JSON-unmarshal the nerdctl version output: %w", err)
-	}
-
-	tmpl, err := newVersionTemplate(format)
-	if err != nil {
-		return err
-	}
-	err = va.showVersionMessage(tmpl, nerdctlVersion)
-	if err != nil {
 		return err
 	}
