+++ pkg/config/config_native.go
@@ -1,0 +1,10 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+//go:build linux
+
+package config
+
+type Finch struct {
+	GeneralSettings
+}
