+++ pkg/config/nerdctl_config_applier_test.go
@@ -1,6 +1,8 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+//go:build (darwin || windows) && !native
+
 package config
 
 import (
