+++ pkg/config/validate_natvie.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build windows && !native
+//go:build linux || native
 
 package config
